+++ compiler-tests/src/test/data/diagnostic/aggregation/SimilarBindingsNotReportedWhenMissingBindingTypeIsAny.kt
@@ -1,0 +1,11 @@
+// RUN_PIPELINE_TILL: FIR2IR
+// RENDER_IR_DIAGNOSTICS_FULL_TEXT
+
+@DependencyGraph(AppScope::class)
+interface AppGraph {
+  val <!METRO_ERROR!>any<!>: Any
+
+  @Provides fun int(): Int = 3
+  @Provides fun str(): String = "string"
+  @Provides fun boolean(): Boolean = true
+}